Move files into packages
