+++ src/main/java/lemon/app/Launcher.java
@@ -1,6 +1,8 @@
-package lemon;
+package lemon.app;
 
 import javafx.application.Application;
+import lemon.ui.LemonFx;
+
 /**
  * A launcher class to workaround classpath issues.
  */
