+++ src/main/java/lemon/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package lemon;
+package lemon.ui;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -26,7 +26,7 @@ public class DialogBox extends HBox {
 
     private DialogBox(String text, Image img) {
         try {
-            FXMLLoader fxmlLoader = new FXMLLoader(lemon.LemonFxWindow.class.getResource("/view/DialogBox.fxml"));
+            FXMLLoader fxmlLoader = new FXMLLoader(LemonFxWindow.class.getResource("/view/DialogBox.fxml"));
             fxmlLoader.setController(this);
             fxmlLoader.setRoot(this);
             fxmlLoader.load();
