+++ src/main/java/lemon/ui/LemonFx.java
@@ -1,4 +1,4 @@
-package lemon;
+package lemon.ui;
 
 import java.io.IOException;
 
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import lemon.Lemon;
 
 /**
  * A GUI for Duke using FXML.
