+++ src/main/java/lemon/ui/LemonFxWindow.java
@@ -1,4 +1,4 @@
-package lemon;
+package lemon.ui;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -7,6 +7,8 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+import lemon.Lemon;
+
 /**
  * Controller for the main GUI.
  */
